Fix crash when share is called from an iPad
It was crashing because iPad have a different requirement for showing
the `UIActivityViewController`. Specifically, they must be presented in
a popover.

Note that, since popovers require to have a point (or view) of origin,
we set it to be the bottom center of the view, mimicking the behaviour
of iPhones.

sources:
- https://developer.apple.com/documentation/uikit/uiactivityviewcontroller
- https://stackoverflow.com/a/33942956/3600120
- https://stackoverflow.com/a/54140310/3600120